rename email.py to email_notifier.py
email.py is also a standard python module, and the same names appear to
confuse python.
